requirements.txt
ReFrame-HPC==4.7.4
